+++ CHANGES.rst
@@ -13,14 +13,18 @@ Changelog of threedi-modelchecker
   instead of warning.
 - Rewrite command-line client. The `--sqlite` argument is now an argument of the
   `check` command, not of the main `threedi_modelchecker` group. To run a check,
-  the new syntax is 
+  the new syntax is
+
   `threedi_modelchecker check -s <your database>.sqlite -l <desired check level>`
+
   Run `threedi_modelcheck check --help` for an overview.
 
   A new command, `export-checks`, has also been added. This exports all checks
   executed by the model checker as an RsT table or in CSV format, as specified by
   the optional `--format` argument. The check output can also be dumped to a file
-  using `--file`. Run `threedi_modelcheck export-checks --help` for an overview.
+  using `--file`.
+  
+  Run `threedi_modelcheck export-checks --help` for an overview.
 
 
 1.0.1 (2023-02-02)