+++ CHANGES.rst
@@ -29,7 +29,7 @@ Changelog of threedi-modelchecker
 - Add check 326: this gives an info message if a record exists in the simple_infiltration
   table, but is not referenced from the global settings.
 
-- Modify check 1202 to also check for an empty string as a timeseries.
+- Add check 1205 to make sure that a timeseries is not an empty string.
 
 
 1.0.1 (2023-02-02)